LayoutTests/platform/ios-simulator-18/imported/w3c/web-platform-tests/uievents/mouse/offset_and_page_after_dispatch.tentative-expected.txt


FAIL Click on the target without moving the mouse cursor assert_equals: mousedown should've been received expected (string) "mousedown" but got (undefined) undefined
